Made RemoteObject derive Clone and increased size of timestamp in RemoteObject

+++ src/domain/runtime/event/mod.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 use std::fmt;
 pub mod console_api_called;
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 #[non_exhaustive]
 #[serde(tag = "method", content = "params")]
 pub enum Event {
@@ -15,15 +15,15 @@ pub enum Event {
     ExceptionThrown(ExceptionParams),
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 #[non_exhaustive]
 #[serde(rename_all = "camelCase")]
 pub struct ExceptionParams {
-    pub timestamp: i32,
+    pub timestamp: i64,
     pub exception_details: ExceptionDetails,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct ExceptionDetails {
     pub text: String,
